Now input box can be selected and unselected for typing Added arrow drawing functionality though not tested

+++ src/Menu.py
@@ -33,15 +33,21 @@ def menu():
                 py.quit()
                 sys.exit()
 
-            if join_button.is_over(py.mouse) and event.type == py.MOUSEBUTTONDOWN:
-                join_button.actions[0] = join(ip_input.text.text)
-                join_button.exe_all()
+            if event.type == py.MOUSEBUTTONDOWN:
+                if join_button.is_over(py.mouse):
+                    join_button.actions[0] = join(ip_input.text.text)
+                    join_button.exe_all()
 
-            elif host_button.is_over(py.mouse) and event.type == py.MOUSEBUTTONDOWN:
-                host_button.actions[0] = host()
-                host_button.exe_all()
+                elif host_button.is_over(py.mouse):
+                    host_button.actions[0] = host()
+                    host_button.exe_all()
 
-            if event.type == py.KEYDOWN:
+                elif ip_input.is_over(py.mouse):
+                    ip_input.select()
+                else:
+                    ip_input.unselect()
+
+            if event.type == py.KEYDOWN and ip_input.active:
                 if event.key == py.K_BACKSPACE:
                     ip_input.delete()
                 else:
